Rename tests crate to north_tests
