+++ .cargo/linux_runner_wrapper.sh
@@ -2,7 +2,7 @@
 
 binary_name=`basename $1`
 
-if [[ $binary_name =~ ^(north|tests-[a-z0-9]{16})$ ]]; then
+if [[ $binary_name =~ ^(north|integration_tests-[a-z0-9]{16})$ ]]; then
     sudo -E $@
 else
     eval $@
