+++ Cargo.toml
@@ -7,8 +7,8 @@ members = [
     "npk",
     "nstar",
     "sextant",
-    "tests",
-    "tests/test_container",
+    "north_tests",
+    "north_tests/test_container",
     "examples/container/cpueater",
     "examples/container/crashing",
     "examples/container/datarw",
