+++ north_tests/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "tests"
+name = "north_tests"
 version = "0.1.0"
 authors = ["ESRLabs"]
 edition = "2018"
@@ -10,5 +10,6 @@ async-trait = "0.1.36"
 env_logger = "0.7.1"
 log = "0.4.8"
 north = { path = "../north" }
+nstar = { path = "../nstar" }
 regex = "1.3.9"
 tokio = { version = "0.3", features = ["full"] }
